Refactor views to seperate device, cable, interface, locations modules
